requirements.txt
line-bot-sdk
flask
beautifulsoup4
urlopen
requests

